in the middle of the way
