+++ web/broker/models.py
@@ -15,15 +15,24 @@ class ApplicationForm(models.Model):
         return ApplicationResponse.objects.filter(answers__question__form=self)
 
 class Question(models.Model):
+    class Meta:
+        constraints = [models.UniqueConstraint(fields=['form', 'number'], name='unique_form_number')]
+
     form = models.ForeignKey(ApplicationForm, on_delete=models.CASCADE, related_name="questions")
     question = models.CharField("question", max_length=QUESTION_MAX_LENGTH)
     number = models.IntegerField()
 
+    def typed(self):
+        print(self.__class__)
+        return 0
+        #self.__getattribute__(self.__class__)
+
+    def make_answer(self):
+        raise NotImplementedError
+
     def __str__(self):
         return self.question
 
-    class Meta:
-        constraints = [models.UniqueConstraint(fields=['form', 'number'], name='unique_form_number')]
 
 
 class ApplicationResponse(models.Model):
@@ -42,15 +51,21 @@ class Answer(models.Model):
     question = models.ForeignKey(Question, on_delete=models.CASCADE, related_name='answers')
 
 class TextualQuestion(Question):
-    pass
-class TextAnswer(Answer):
+    def make_answer(self):
+        t = TextualAnswer()
+        t.question = self
+
+class TextualAnswer(Answer):
 
     value = models.CharField('text_value', max_length=100, default="kooooon")
     def __str__(self):
         return self.value
 
 class MultiChoiceQuestion(Question):
-    pass
+    def make_answer(self):
+        t = MultiChoiceAnswer()
+        t.question = self
+
 class MultiChoiceAnswer(Answer):
     value = models.CharField("choice_value", max_length=CHOICE_MAX_LENGTH)
 
@@ -59,7 +74,10 @@ class MultiChoiceAnswer(Answer):
 
 
 class NumericalQuestion(Question):
-    pass
+    def make_answer(self):
+        t = NumericalAnswer()
+        t.question = self
+
 class NumericalAnswer(Answer):
     value = models.IntegerField('int_value')
