+++ web/broker/oldshits.py.bak
@@ -1,0 +1,21 @@
+
+from django.shortcuts import render, redirect
+from django.db.models.query import EmptyQuerySet
+
+from .models import *
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
+
+
+def form_filling(request):
+    return render(request, 'broker')
+
+def display(request):
+    return HttpResponse('See Asghar : %s' % display_form(ApplicationResponse.objects.first()))
+
+
+def display_form(res: ApplicationResponse):
+    html = ""
+    for a in res.answers.order_by('question__number'):
+        html += "<p> {} </p>".format(str(a.question)) + "<p>{}</p>".format(str((TextAnswer)(a)))
+    return html
