+++ web/broker/templates/broker/instructor/home.html
@@ -4,49 +4,38 @@
 {% block content %}
   <div class="ui main text container" style="padding-left: -10em; ">
     <h1 class="ui header">List of Applications</h1>
-      <div class="ui one column stackable grid">
+    {% for form in forms %}
+      <div class="ui raised segment">
+          <div class="ui equal width grid">
+              <div class="column middle aligned center aligned">
+                  <p>{{ form.course_id }}</p>
+              </div>
 
+              <div class="column middle aligned center aligned">
+                  <div class="ui label">
+                    Deadline
+                    <div class="detail">{{ form.deadline|date:"M d" }}</div>
+                  </div>
+              </div>
 
-        <div class="column">
-          {% for form in forms %}
-            <div class="ui raised segment">
-                <div class="ui five column stackable grid">
+              <div class="column middle aligned center aligned">
+                  <div class="ui label">
+                    Responses
+                    <div class="detail">{{ form.get_responses.count }}</div>
+                  </div>
+              </div>
 
-                    <div class="column middle aligned center aligned">
-                        <p>{{ form.course_id }}</p>
+              <div class="column middle aligned center aligned">
+                  <a class="ui animated grey button" tabindex="0" href="{% url 'form_detail' id=form.id %}">
+                    <div class="visible content">Details</div>
+                    <div class="hidden content">
+                      <i class="right arrow icon"></i>
                     </div>
+                </a>
+              </div>
 
-                    <div class="column middle aligned center aligned">
-                        <div class="ui label">
-                              <i class="clock outline icon"></i> {{ form.release_date }}
-                        </div>
-                    </div>
-
-                    <div class="column middle aligned center aligned">
-                        <div class="ui label">
-                              <i class="clock icon"></i> {{ form.deadline }}
-                        </div>
-                    </div>
-
-                    <div class="column middle aligned center aligned">
-                        <div class="ui label">
-                              <i class="pen square icon"></i> {{ form.get_responses.count }}
-                        </div>
-                    </div>
-
-                    <div class="column middle aligned center aligned">
-                        <a class="ui animated grey button" tabindex="0" href="{% url 'form_detail' id=form.id %}">
-                          <div class="visible content">Details</div>
-                          <div class="hidden content">
-                            <i class="right arrow icon"></i>
-                          </div>
-                      </a>
-                    </div>
-
-                </div>
-            </div>
-          {% endfor %}
+          </div>
       </div>
-    </div>
+    {% endfor %}
   </div>
 {% endblock %}
