+++ web/broker/templates/broker/student/application.html
@@ -1,0 +1,4 @@
+{% extends "broker/base.html"%}
+{% block content %}
+{{ form }}
+{% endblock %}
