+++ web/taremi/settings.py
@@ -136,4 +136,6 @@ STATICFILES_DIRS = [
     os.path.join(BASE_DIR, "static"),
 ]
 
+LOGIN_URL = '/account/login/'
+
 LOGIN_REDIRECT_URL='/'
